promoted qm work, added extra tests for accessor with policy
