+++ test/test.cbp
@@ -12,7 +12,7 @@
 				<Option object_output="obj/Debug/" />
 				<Option type="1" />
 				<Option compiler="gcc" />
-				<Option parameters="-nv -tc=compute_weighted_quantiles" />
+				<Option parameters="-nv -tc=accessor_policy" />
 				<Compiler>
 					<Add option="-g" />
 				</Compiler>
@@ -36,7 +36,7 @@
 					<Add directory="../bin/Release" />
 				</Linker>
 				<ExtraCommands>
-					<Add after="cd ../bin/Release &amp;&amp; ./test_shyft -nv  -tc=compute_weighted_quantiles" />
+					<Add after="cd ../bin/Release &amp;&amp; ./test_shyft -nv  -tc=accessor_policy" />
 					<Mode after="always" />
 				</ExtraCommands>
 			</Target>
@@ -146,7 +146,9 @@
 		<Unit filename="pt_ss_k_test.cpp">
 			<Option virtualFolder="method_stacks/" />
 		</Unit>
-		<Unit filename="qm_test.cpp" />
+		<Unit filename="qm_test.cpp">
+			<Option virtualFolder="time_series/" />
+		</Unit>
 		<Unit filename="region_model_test.cpp">
 			<Option virtualFolder="full_model/" />
 		</Unit>
